refactor: Date 검증 및 변환 로직을 utils로 분리 #54
