+++ back/src/main/java/com/developaw/harupuppy/domain/schedule/dto/request/ScheduleCreateRequest.java
@@ -5,6 +5,7 @@ import com.developaw.harupuppy.domain.schedule.domain.RepeatType;
 import com.developaw.harupuppy.domain.schedule.domain.Schedule;
 import com.developaw.harupuppy.domain.schedule.domain.ScheduleType;
 import com.developaw.harupuppy.domain.user.dto.UserScheduleDto;
+import com.developaw.harupuppy.global.utils.DateUtils;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
@@ -26,7 +27,7 @@ public record ScheduleCreateRequest(
     public static Schedule fromDto(ScheduleCreateRequest dto, String homeId, String repeatId) {
         validateDateTime(dto.scheduleDate, dto.scheduleTime);
         return Schedule.builder()
-                .scheduleDateTime(Schedule.parseDateTime(dto.scheduleDate(), dto.scheduleTime()))
+                .scheduleDateTime(DateUtils.parseDateTime(dto.scheduleDate(), dto.scheduleTime()))
                 .scheduleType(dto.scheduleType())
                 .homeId(homeId)
                 .mates(new ArrayList<>())
@@ -40,7 +41,7 @@ public record ScheduleCreateRequest(
     public static Schedule fromDto(ScheduleCreateRequest dto, String homeId) {
         validateDateTime(dto.scheduleDate, dto.scheduleTime);
         return Schedule.builder()
-                .scheduleDateTime(Schedule.parseDateTime(dto.scheduleDate(), dto.scheduleTime()))
+                .scheduleDateTime(DateUtils.parseDateTime(dto.scheduleDate(), dto.scheduleTime()))
                 .scheduleType(dto.scheduleType())
                 .homeId(homeId)
                 .mates(new ArrayList<>())
