+++ back/src/main/java/com/developaw/harupuppy/global/utils/DateUtils.java
@@ -1,0 +1,33 @@
+package com.developaw.harupuppy.global.utils;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
+
+public class DateUtils {
+    static final DateTimeFormatter dateFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+    static final DateTimeFormatter timeFormatter = DateTimeFormatter.ofPattern("HH:mm");
+
+    public static void validateDateTime(String date, String time) {
+        try {
+            LocalDate.parse(date, DateTimeFormatter.ofPattern("yyyy-MM-dd"));
+            LocalTime.parse(time, DateTimeFormatter.ofPattern("HH:mm"));
+        } catch (Exception e) {
+            throw new IllegalArgumentException("스케줄 날짜와 시간이 유효하지 않습니다", e);
+        }
+    }
+    public static void validateDate(String date) {
+        try {
+            LocalDate.parse(date, DateTimeFormatter.ofPattern("yyyy-MM-dd"));
+        } catch (Exception e) {
+            throw new IllegalArgumentException("생일 날짜가 유효하지 않습니다", e);
+        }
+    }
+
+    public static LocalDateTime parseDateTime(String date, String time) {
+        LocalDate localDate = LocalDate.parse(date, dateFormatter);
+        LocalTime localTime = LocalTime.parse(time, timeFormatter);
+        return LocalDateTime.of(localDate, localTime);
+    }
+}
